### 2020-06-20 #2374 @whedon
[PRE REVIEW]: Alarconpy: a Python Package for Meteorologists
**Submitting author:** @apalarcon (<a href="http://orcid.org/0000-0002-9454-2331">Albenis Pérez Alarcón</a>)
**Repository:** <a href="https://github.com/apalarcon/alarconpy/" target ="_blank">https://github.com/apalarcon/alarconpy/</a>
**Version:** v1.0.3
