+++ paper/paper.bib
@@ -27,7 +27,7 @@
             and Goebbert, Kevin and Thielen, Jonathan E. and Bruick, Zachary S.},
   organization = {Unidata},
   title = {MetPy: A {Python} Package for Meteorological Data},
-  year = {2008 - 2020},
+  year = {2020},
   version = {0.12.1.post2},
   doi = {10.5065/D6WW7G29},
   url = {https://github.com/Unidata/MetPy},
@@ -39,7 +39,7 @@
   author = {May, Ryan and Arms, Sean and Leeman, John and Chastang, Julien },
   organization = {Unidata},
   title = {Siphon: A collection of {Python} Utilities for Accessing Remote Atmospheric and Oceanic Datasets},
-  year = {2014 - 2017},
+  year = {2017},
   version = {0.5.1},
   doi = {10.5065/D6CN72NW},
   url = {https://github.com/Unidata/siphon},
