### 2020-06-30 #2374 @tomchor
Sorry, but I am in the process of writing my PhD thesis so I'll have to
pass because I simply will not have time.

Best of luck!
Cheers

On Tue, Jun 30, 2020 at 7:53 AM David H Hagan <notifications@github.com>
wrote:

> @dvalters <https://github.com/dvalters>, @tomchor
> <https://github.com/tomchor>, @hertzsprung
> <https://github.com/hertzsprung>, @arbennett
> <https://github.com/arbennett> - Do any of you have the time to review
> this manuscript?
>
