### 2020-06-30 #2374 @hertzsprung
I'm happy to review, what's the timescale? (first-time JOSS reviewer here)
