### 2020-06-30 #2374 @dhhagan
Hey, @hertzsprung welcome! There is no hard deadline that I'm aware of, though I will double-check. I would say on the order of 2-4 weeks?

@arfon Is there a guideline on the appropriate length of time for a review to take place?
