### 2020-06-30 #2374 @danielskatz
> Is there a guideline on the appropriate length of time for a review to take place?

In normal times, 2-3 weeks, but longer is ok if needed due to covid-related things
