### 2020-07-13 #2374 @RCHG
I could review the manuscript but it will be my first time on JOSS so I have to learn about the process.
