### 2020-07-19 #2479 @dhhagan
Hi @RCHG - you should be able to edit the issue once you accept the invite. If you did not get an invite, please let me know and I will re-invite you to the issue. As for (b) - yes, please indicate the steps right here in this issue. Please let me know if you have any other qustions!
