### 2020-07-23 #2479 @dhhagan
Hi @hertzsprung - please resend! I haven't seen it yet, but am happy to answer any questions. Feel free to ask here in the comments as well if you feel comfortable doing so.
