### 2020-08-05 #2479 @hertzsprung
In my view, Alarconpy fails to represent a substantial scholarly effort because it comprises numerous utility functions that are already provided by existing libraries, including:

* interpolation functions already provided by scipy
* statistics functions and haversine function already provided by scikit-learn
* unit conversion already provided by [Unum](https://github.com/trzemecki/Unum)

It would be more appropriate for any missing features to be added to these existing libraries.  Some of Alarconpy's thermodynamic functions might find a good home in [metpy](https://unidata.github.io/MetPy/), which is already a project dependency.
