### 2020-08-07 3f53f0c
Update plots_hurricanes.py
