### 2021-01-05 #2479 @kgoebber
**General Comments**

Overall this package provides a mechanism to facilitate the inspection of various meteorological data. Most advanced functionality in this package comes from other well-established Python modules including SciPy and MetPy. Overall, it is hard to see extensive use of this package by the broader community due its limited scope, scant documentation, and better offering of similar functionality in other Python libraries.

**Specific Comments**
Substantial scholarly effort: This package does not meet the criteria for acceptance. It is a young project with a relatively few number of commits. This effort is focused on the main authors work flow and has a couple of features not present in other modules, but would be better served by inclusion in those packages.

Installation: Installation went fine, but it was confusing which folder to copy as there as both the main repository folder and the source code folder are named the same. Dependency installation is left to the user, which is not as problematic in this instance due to all required packages are easy installed through package management systems.

Functionality: In general, what I tested worked fine, but it was hard to determine the full extent of the functionality through the limited documentation and examples given in the repository.

A statement of need: While there was some basic statement of how this package might be useful to others, the scope of the project is not beyond that for which other packages already exist within the domain space. For example, the plotting capabilities of this repository make getting basic plots relatively easy, there is more robust functionality for similar plots already in MetPy that make full use of that library.

Example usage:  There are a number of examples, however, some of them are incomplete. For example, the interpolation example leaves out the important data input lines to have a clear example of what data are going into the interpolation routines.

Automated tests: No tests are conducted as a part of this package to verify accuracy and consistency as changes are made.

Community guidelines: There are no guidelines for being able to contribute to this package, however, there is information on who to contact, via email, about the package and its operation.

Software paper: The paper fulfills the basic requirements set forth in a statement of need and context of this work within the Python ecosystem.
