### 2021-01-11 #2479 @dhhagan
@apalarcon @hertzsprung @kgoebber @openjournals/jose-eics

First, I would like to thank you (@apalarcon) for submitting to JOSS. I hope that, while delayed, this process was useful and helpful for you and your work. Upon review, both reviewers have recommended this manuscript for rejection from JOSS as it currently stands due to the fact it mostly represents smaller utility functions than/should be incorporated into existing packages where they aren't already and the supporting work around the main code needs to be substantially improved (see specific comments from @kgoebber above). Thank you very much for taking the time to submit and go through the review process with JOSS - I will let the @openjournals/jose-eics take it over from here. Please reach out if you have any questions, comments, or concerns.
