### 2021-02-02 #2479 @kyleniemeyer
Hi @apalarcon, my apologies but I need one more edit to the paper from you: can you please add a short Statement of Need section? This is part of our [required elements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) for a JOSS paper. You may be able to move some of the text in your summary.

There is no need to re-deposit the software in Zenodo, since JOSS archives the paper text.

(Normally whedon checks for this section upon submission, I'm not sure how we missed this.)
