+++ README.md
@@ -1,7 +1,5 @@
 # Alarconpy: A Python Package for Meteorologists
 
-
-Markdown:
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.02479/status.svg)](https://doi.org/10.21105/joss.02479)
 
 
