### 2021-02-03 #2479 @arfon
Hi all, it looks like a mistake was made here in the handoff between editors on rotation this week. For context, it looks like the following happened:

- @dhhagan (the handling editor) reminded me on Slack about this paper. I failed to notice that this was a paper that we had been discussing before and so asked him to run `@whedon accept` which he did [here](https://github.com/openjournals/joss-reviews/issues/2479#issuecomment-770405526). This is how editors typically notify the EiC team that the paper is ready to publish and accept (which is incorrect in this case).
- The newly incoming EiC on rotation appears to have noticed the `@whedon accept`, and presumably thinking the paper was in good shape to move forward, asked the author to make an archive of their software so that they could continue the acceptance process.

Based on the thorough feedback of the reviewers, our decision here was actually to _reject_ this submission as it fails to meet our requirements for substantial scholarly effort.

I'll work with the editorial team to work out what our next steps here and report back later today on how we plan to handle this submission.

Apologies to all for any confusion and frustration caused. This looks to simply be an error on our part.
