### 2021-02-03 #2479 @arfon
Hi all, just a follow-up on my earlier message here. I've confirmed the above events with the EiC team.

As the acceptance here was processed in error, we've contacted Crossref (the DOI registering service for JOSS) who have deleted the DOI associated with this paper (https://doi.org/10.21105/joss.02479).

@apalarcon, @dhhagan, @hertzsprung, @kgoebber - my sincere apologies for any confusion caused here. We will use this incident to learn and improve our processes.

@apalarcon - based on the feedback of both reviewers and in discussions with the handling editor here, the decision is to reject this submission as it fails to meet our requirements for substantial scholarly effort.

Re-submissions are welcome in JOSS, but as @dhhagan mentioned outlined above, this submission would require significant work to be considered publishable here.
