### 2021-02-17 #2479 @dopplershift
@arfon Just an FYI, the PDF is still available [here](https://www.theoj.org/joss-papers/joss.02479/10.21105.joss.02479.pdf), which is being picked up by Google Scholar.
